[MISC] Move include/hibf/detail/layout/ -> include/hibf/layout
