+++ include/hibf/detail/data_store.hpp
@@ -8,7 +8,7 @@
 #include <string>    // for string
 #include <vector>    // for vector
 
-#include <hibf/detail/layout/layout.hpp>      // for layout
+#include <hibf/layout/layout.hpp>      // for layout
 #include <hibf/detail/sketch/hyperloglog.hpp> // for hyperloglog
 
 namespace seqan::hibf
