+++ include/hibf/layout/hierarchical_binning.hpp
