+++ include/hibf/layout/simple_binning.hpp
