+++ src/layout/simple_binning.cpp
@@ -5,8 +5,8 @@
 #include <vector>    // for vector
 
 #include <hibf/detail/data_store.hpp>            // for data_store
-#include <hibf/detail/layout/layout.hpp>         // for layout
-#include <hibf/detail/layout/simple_binning.hpp> // for simple_binning
+#include <hibf/layout/layout.hpp>         // for layout
+#include <hibf/layout/simple_binning.hpp> // for simple_binning
 
 namespace seqan::hibf::layout
 {
