+++ test/unit/hibf/detail/layout/hierarchical_binning_test.cpp
@@ -5,9 +5,9 @@
 
 #include <hibf/config.hpp>                               // for config
 #include <hibf/detail/data_store.hpp>                    // for data_store
-#include <hibf/detail/layout/compute_fpr_correction.hpp> // for compute_fpr_correction
-#include <hibf/detail/layout/hierarchical_binning.hpp>   // for hierarchical_binning
-#include <hibf/detail/layout/layout.hpp>                 // for layout
+#include <hibf/layout/compute_fpr_correction.hpp> // for compute_fpr_correction
+#include <hibf/layout/hierarchical_binning.hpp>   // for hierarchical_binning
+#include <hibf/layout/layout.hpp>                 // for layout
 #include <hibf/test/expect_range_eq.hpp>                 // for expect_range_eq, EXPECT_RANGE_EQ
 
 TEST(hierarchical_binning_test, small_example)
