+++ util/fpr_quality.cpp
@@ -10,7 +10,7 @@
 
 #include <hibf/contrib/robin_hood.hpp>
 #include <hibf/detail/build/bin_size_in_bits.hpp>
-#include <hibf/detail/layout/compute_fpr_correction.hpp>
+#include <hibf/layout/compute_fpr_correction.hpp>
 #include <hibf/interleaved_bloom_filter.hpp>
 
 #include <sharg/parser.hpp>
